MDEV-8565: COLUMN_CHECK fails on valid data
When an INTEGER type dynamic column has a value of ZERO, it is stored as a zero-length amount of data (nothing in the data section at all). This change now ignores the data length in COLUMN_CHECK for INTEGER data types, since zero-length data is a valid result produced by COLUMN_CREATE due to the way variable length INTEGER data is stored.