Fix segfault with musl libc due to bad use of strerror_r
The only known implementation that has the GNU variand of strerror_r is
GNU libc. Building with musl libc and _GNU_SOURCE enables some other gnu
extensions, but musl libc does not implement the broken strerror_r.

We fix this by check explicitly for GNU libc in addition to _GNU_SOURCE
and if they both are not set, then fall back to standard.

This fixes segfault with musl libc.