MDEV-9095: report successs of mlock as warning
Failure attempts where logged but its important to report success too
otherwise a user may be left with the impression that a non-root user
attempt at mlock didn't succeed with the inference that it wasn't tried
(successfully) as a root user.

Attempt to raise effective id to root regardsless of non-root failure
reason. i.e. ENOMEM because of MEMLOCK resource limit or lack of
CAP_IPC_LOCK capability.

Remove locked_in_memory=0 as in else branch now this is the only thing
it can be.