rsync_sst_rsync: ignore lost+found during database directory transfer
lost+found is excluded when transferring normal directories. The
database directory transfer however does not exclude it. As a result, wsrep_sst_rsync in
donor mode attempts  to read its content. This fails as it is owned by
root with strict 700 permissions.

This change excludes lost+found from the list of database directories in
the wsrep_sst_rsync script.