remove die() from BUILD/autorun.sh
This patch removes die() function from the BUILD/autorun.sh. It was
introduced in the c682570431 commit (Fix BUILD/autorun.sh to really bail
out on error.). Last users of die() was removed in the 8664de22 commit
(WL#5665: Removal of the autotools-based build system) and since it is
not used anywhere.

No functionality changes. Just cleanup.