Fix assertion/hang in read_init_file()
If there are other threads running (for example binlog background
thread), then the thread count may not drop to zero at the end of
do_handle_bootstrap(). This caused an assertion and missing wakeup of
the main thread. The missing wakeup is because THD::~THD() only
signals the COND_thread_count mutex when the number of threads drops
to zero.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>