Make sure to recompile the feedback plugin for EMBEDDED
On OS X with clang 7.3.0 running any test as embedded crashes. As
libfeedback was not compiled with embedded, we had 2 notions of THD,
with different number of members. When calling inline functions defined in
sql_class.h, the functions were defined for 1 THD but called on an
object that is different. This resulted in erroneous data being returned
and shortly after, a crash within the alloc_query() function.

Recompile the feedback plugin for the embedded server specifically to
not have such symbol conflicts.