MDEV-11082 mysql_client_test: test_ps_query_cache fails with group-concat-max-len=1M
test_bug14169 was setting session group_concat_max_len=1024 and
did not clean it up. Because of that test_ps_query_cache, when run
with group-concat-max-len != 1024, had different values in connections,
and was inserting into query cache when a hit was expected.

Fixed by adding a clean-up for the value in test_bug14169